Rename packages for archiver code
* Move exploratory tests into separate folder. They are only used for experimenting locally.
