+++ status-service/archiverv1/conversion_test.go
@@ -1,4 +1,4 @@
-package oldarchiver
+package archiverv1
 
 import (
 	"testing"
