+++ status-service/archiverv2/client.go
@@ -1,4 +1,4 @@
-package archiver
+package archiverv2
 
 import (
 	"context"
